requirements.txt
streamlit
tensorflow==2.16.1  # Explicit version for compatibility
numpy==1.26.4
Pillow==10.3.0
h5py==3.11.0